Update to latest GENESIS version
I had an older version of GENESIS installed, so the null model was
in the old format. Update the version and rerun the data-raw null
model script to get a null model in the new version. Save the old
version as its own file for testing. Make a new renv snapshot to
reflect the GENESIS update.