Specify boxmode to fix grouped boxplots
Filled boxplots did not have position = dodge in plotly - instead,
all boxes appear on top of each other. This can be fixed by setting
the layout of boxmode to group, but it also produces a number of
warnings. These warnings can apparently be ignored, but I haven't
found a good way to do that yet. (suppressWarnings seems dangerous,
and expect_warning makes the test code cluttered.)

Info about this warning:
https://github.com/ropensci/plotly/issues/1644
https://community.plotly.com/t/not-a-real-error-for-boxmode/2596