Add basic test for overview module
This module tests whether the overview module loads. This module
doesn't do much other than display information and show an action
button that switches tabs. Because we just want to check if the
module runs and the button exists, this test only sets the input
button and then tests a condition that is always true.