Use input instead of reactive to determine plot type
Remove the reactive plot_type(), and use the input plot_type to
determine the plot type. Remove the observe statement that updates
the text on the action button - it now just says "Generate plot".
Remove the output plot_type - it is no longer needed since the user
selects it as an input. Update conditional panel conditions to use
the input plot_type instead of the output.