Add unit test to replicate data loading bug
The Load data tab fails when only a null model and a phenotype
file are specified, which should be allowed because the genotype
file is optional. This likely occurred because I don't have good
unit tests for the shinyFileChoose functionality (and couldn't
figure out how to write them), so this behavior was unfortunately
not tested. Note that the shinyFileChoose functionality will be
replaced anyway, so this is ok for now. In any case, add a test to
check that passing a 0-length character vector to the .load_data
function returns the correct data.